[#283] refactor: template으로 패키지 및 클래스 이름 변경 - 기존 전략 두 개 혼용으로 알 수 없어 명확하게 구분 시 - BadgeAwardStrategyTemplate -> BadgeAwardTemplate - 패키지 strategytemplate -> template
