+++ src/main/java/goorm/eagle7/stelligence/domain/badge/template/impl/ContributeRejectedBadgeAwardTemplate.java
@@ -1,4 +1,4 @@
-package goorm.eagle7.stelligence.domain.badge.strategytemplate.impl;
+package goorm.eagle7.stelligence.domain.badge.template.impl;
 
 import static goorm.eagle7.stelligence.domain.badge.model.Badge.*;
 
@@ -9,38 +9,36 @@ import org.springframework.stereotype.Component;
 
 import goorm.eagle7.stelligence.domain.badge.model.Badge;
 import goorm.eagle7.stelligence.domain.badge.model.BadgeCategory;
-import goorm.eagle7.stelligence.domain.badge.strategytemplate.BadgeAwardStrategyTemplate;
+import goorm.eagle7.stelligence.domain.badge.template.BadgeAwardTemplate;
 import goorm.eagle7.stelligence.domain.contribute.ContributeRepository;
 import goorm.eagle7.stelligence.domain.contribute.model.ContributeStatus;
 import lombok.RequiredArgsConstructor;
 
 @Component
 @RequiredArgsConstructor
-public class ContributeMergedBadgeAwardStrategyTemplate extends BadgeAwardStrategyTemplate {
+public class ContributeRejectedBadgeAwardTemplate extends
 
-	private static final Map<Integer, Badge> requiredCounts = new HashMap<>();
+	BadgeAwardTemplate {
+
 	private final ContributeRepository contributeRepository;
+	private static final Map<Integer, Badge> requiredCounts = new HashMap<>();
 
 	@Override
 	public boolean supports(BadgeCategory category) {
-		return BadgeCategory.CONTRIBUTE_MERGED.equals(category);
+		return BadgeCategory.CONTRIBUTE_REJECTED.equals(category);
 	}
 
 	@Override
 	protected long getCount(Long memberId) {
-		return contributeRepository.countByMemberIdAndStatus(memberId, ContributeStatus.MERGED);
+		return contributeRepository.countByMemberIdAndStatus(memberId, ContributeStatus.REJECTED);
 	}
 
 	@Override
 	protected Map<Integer, Badge> getBadgeCriteria() {
 
 		if (requiredCounts.isEmpty()) {
-			requiredCounts.put(1, JUPITER);
-			requiredCounts.put(5, SATURN);
-			requiredCounts.put(10, PLUTO);
-			requiredCounts.put(50, ANDROMEDA);
+			requiredCounts.put(100, BLACKHOLE);
 		}
 		return requiredCounts;
 	}
-
 }
